Address PR Comments
This commit addresses comments from the open PR. Including:
- Check for nulls in Injected c-tors and other public methods
- Refactor Decoder and Encoder to simplify, avoiding interface and factory
- Provide informative exception messages for unsupported properties
- Ensure Trino codestyle followed (no final local variables)

I also discovered and fixed a bug where null map values were not being
encoded correctly.
